+++ CHANGELOG.md
@@ -11,6 +11,8 @@ _The old changelog can be found in the `release-2.6` branch_
 
 # Changes Since v3.1.0
 
+  - Fixed usage docstrings for OCI - was missing "oci" in command examples.
+
 # v3.1.0 - [2019.02.22]
 
 ## New Commands